chatService 중간 commit

+++ src/main/java/com/tolover/tolover/service/JwtService.java
@@ -153,7 +153,7 @@ public class JwtService {
         }
     }
 
-    public int getUserIdx() throws BaseException{
+    public long getUserIdx() throws BaseException{
         String accessToken = getAccessToken();
 
         if (accessToken == null || accessToken.length() == 0) {
